+++ backend/src/main/java/com/woowacourse/gongseek/member/presentation/dto/MyPageArticleResponse.java
@@ -40,17 +40,4 @@ public class MyPageArticleResponse {
                 article.getViews()
         );
     }
-
-    @Override
-    public String toString() {
-        return "MyPageArticleResponse{" +
-                "id=" + id +
-                ", title='" + title + '\'' +
-                ", category='" + category + '\'' +
-                ", commentCount=" + commentCount +
-                ", createdAt=" + createdAt +
-                ", updatedAt=" + updatedAt +
-                ", views=" + views +
-                '}';
-    }
 }
