+++ backend/src/main/resources/static/docs/index.html
@@ -465,7 +465,7 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2022-08-01 19:47:31 +0900
+Last updated 2022-08-07 16:13:29 +0900
 </div>
 </div>
 </body>
