+++ backend/src/main/resources/static/docs/oauth.html
@@ -489,7 +489,7 @@ Vary: Origin
 Vary: Access-Control-Request-Method
 Vary: Access-Control-Request-Headers
 Content-Type: application/json
-Content-Length: 27
+Content-Length: 25
 
 {
   "url" : "login url"
@@ -532,7 +532,7 @@ Content-Length: 27
 <div class="content">
 <pre class="highlightjs highlight nowrap"><code data-lang="http" class="language-http hljs">POST /api/auth/token HTTP/1.1
 Content-Type: application/json;charset=UTF-8
-Content-Length: 23
+Content-Length: 21
 Host: localhost:8080
 
 {
@@ -571,7 +571,7 @@ Vary: Origin
 Vary: Access-Control-Request-Method
 Vary: Access-Control-Request-Headers
 Content-Type: application/json
-Content-Length: 37
+Content-Length: 35
 
 {
   "accessToken" : "accessToken"
@@ -606,7 +606,7 @@ Content-Length: 37
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2022-07-27 14:26:15 +0900
+Last updated 2022-07-20 11:43:30 +0900
 </div>
 </div>
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.15.6/styles/github.min.css">
